Generalize the starting conditions (failing some unit tests)
In the previous commit, the play always started with N and it was always NS whose trick target mattered.
Some unit tests which fail (GC problem) have been ignored.
There seems to be a sudden explosion in total states as we move from 32 to 36 and then to try to reach 40 levels (see IntegrationTreeSpec).
Some constants replaced by vals (such as CardsPerDeal).
Whist: added new type to represent a whist-type game;'
Deal is now a component of Whist.
AnalyzeDoubleDummy moved from Deal to Whist;